Changed the goal text to "x out of x (A/M/E)" and added a launch screen that i decided to be blank. ALso changed gradesvc to so that it subclasses subjectsvc

+++ NCEACredits/BubbleMainCredits.m
@@ -44,7 +44,7 @@
 }
 
 - (void)setNumberOfCredits:(NSUInteger)credits {
-    _credits.text = [NSString stringWithFormat:@"%i", credits];
+    _credits.text = [NSString stringWithFormat:@"%lu", (unsigned long)credits];
 }
 
 + (NSString *)getCharTitleForGradeType:(NSString *)grade {
